Converts actions to alt style actions
This commit converts most of the actions over to alt style. Since the
actions act as a dispatcher we can just dispatch the payload directly.

One difference of this can be seen in `ChatServerActionCreators`
where we don't separate out the server and view actions using
`handleServerAction` and `handleViewAction`. If you do need to know the
source of an action or wish to have different behavior for an action you
have two options:

Dispatch a payload object with a source parameter.

```js
receiveAll: function(rawMessages) {
  this.dispatch({
    source: 'server-action',
    data: rawMessages
  })
}
```

If you need to have different behavior for server and view actions you
can just implement the behavior in the action itself, since the action
is a dispatcher.