Converts MessageStore
Stores are arguably the most complex piece in flux so we can convert
each in a single commit. We can start with MessageStore, a lot of the
basic internals of the code is left unchanged sans semicolons. The bulk
of the dispatcher logic has been moved inside the store class' action
listener methods that correspond to the proper action. Notably the
underscore variable `_messages` has been introduced as state inside our
store, and all the loose functions that existed are now static functions
of the class.