Converts ThreadStore
Next step is to convert ThreadStore. The main differences are that
currentId and threads are now part of the instance. There is this class
method `_init` which isn't bound to an action, it is keeping in spirit
with the original flux version. Inside `_init` there is a call to
`this.getInstance()` what that does is it gives you the store's instance
so you can call the static methods or use any of the store instance's
functions. The static methods then have access to `this.getState()`
which they use to return the correct data.