Warn and safely continue when store has same name
Previously a ReferenceError was thrown when a store with the same name
was encountered. While this is great, it can be surprising in an
environment where source is minified and this is an error that may be
caught a little too late. With this behavior the store continues working
but logs a warning to the console.