Set state in mixinContainer after mounting
This works around an edge case when a descendent of an AltContainer
updates data in a store on componentDidMount. Without this change, the
descendent updates the flux store after mounting, then the AltContainer
subscribes to the store but doesn't get any updated data between the
initial state and subsequent store changes.