Fixing issues with babel core with a fresh install and test failures
Installed a newer version of babel to fix a path undefined error when running babel src.
The issue was identified and resolved in the newer versions of babel core.

Updated failing tests for batching actions. For the test cases the button click
action was not dispatching anything and only setting a timeout.
Instead just made it return a null when the actions are called so it will
dispatch a null when those actions are called in addition to dispatching.
a action when the timeout runs.